requirements.txt
absl-py==0.13.0
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
debugpy==1.4.1
decorator==5.0.9
gast==0.3.3
google-auth==1.33.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.38.1
h5py==2.10.0
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
Keras-Preprocessing==1.1.2
Markdown==3.3.4
matplotlib-inline==0.1.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.1.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.16.0
SQLAlchemy==1.4.22
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
XlsxWriter==1.4.4
